refactor: update import path for Profile component in App.tsx
